:truck: create api folder for better structure
