+++ src/app/utils/api/index.ts
@@ -1,0 +1,29 @@
+const { VITE_API_ENDPOINT } = import.meta.env;
+
+if (!VITE_API_ENDPOINT) {
+  throw new Error('VITE_API_ENDPOINT is not set');
+}
+
+export function fetchJSON<T>(
+  url: RequestInfo,
+  init?: RequestInit | undefined
+): Promise<T> {
+  return fetch(`${VITE_API_ENDPOINT}${url}`, init).then((response) =>
+    response.json()
+  );
+}
+
+export function createPostOptions(data: BodyInit): RequestInit {
+  const postOptions = {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: data,
+  };
+
+  return postOptions;
+}
+
+export * from './highlights';
+export * from './matches';
